requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==3.2.9
django-cors-headers==3.12.0
django-filter==21.1
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.11.4
oauthlib==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tweepy==4.10.0
urllib3==1.26.9
zipp==3.8.0
